Move the kustomize gitops plugin code to the appropriate file
Follow-up for #973
